Move reducers, fix broken paths
