+++ static/src/containers/Settings/actions.js
@@ -1,4 +1,4 @@
-import makeActionCreator from './makeActionCreator';
+import makeActionCreator from '../../utils/makeActionCreator';
 
 export const TOGGLE_PLAY_ON_HOVER = 'TOGGLE_PLAY_ON_HOVER';
