+++ static/src/containers/Sounds/reducer.js
@@ -1,9 +1,9 @@
 import { FETCH_SOUNDS_SUCCESS, UPDATE_SOUNDS_POSITION, UPDATE_MAP_POSITION,
   SELECT_SOUND_BY_ID, GET_SOUND_BUFFER, TOGGLE_HOVERING_SOUND, PLAY_AUDIO_SRC,
   STOP_AUDIO_SRC, MAP_COMPUTATION_COMPLETE, REMOVE_SOUND }
-  from '../actions/actionTypes';
-import { MAP_SCALE_FACTOR } from '../constants';
-import sessions from './sessions';
+  from './actions';
+import { MAP_SCALE_FACTOR } from '../../constants';
+import sessions from '../Sessions/reducer';
 
 export const computeSoundGlobalPosition = (tsnePosition, spacePosition, mapPosition) => {
   const { translateX, translateY, scale } = mapPosition;
