--- static/src/reducers/sounds.test.js
@@ -1,3 +1,0 @@
-import expect from 'expect';
-import deepFreeze from 'deep-freeze';
-import '../polyfills/Object.assign';
