--- static/src/reducers/spaces.test.js
@@ -1,4 +1,0 @@
-import expect from 'expect';
-import deepFreeze from 'deep-freeze';
-import '../polyfills/Object.assign';
-import { default as reducer } from './spaces';
