--- static/src/store/configureStore.prod.js
@@ -1,11 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import fseReducer from '../reducers';
-
-export default function configureStore(initialState) {
-  return createStore(
-    fseReducer,
-    initialState,
-    applyMiddleware(thunk)
-  );
-}
